Add mapping for player monster, and db monster, also refactor code.

+++ src/main/java/com/gnome/gnome/continueGame/component/Coin.java
@@ -27,7 +27,7 @@ public class Coin {
     /**
      * The value of the coin (amount added to player's coin count).
      */
-    private final int value;
+    private final double value;
     /**
      * The image used to visually represent the coin in the UI.
      */
@@ -40,7 +40,7 @@ public class Coin {
      * @param gridY the Y coordinate on the map grid
      * @param value the numeric value of the coin
      */
-    public Coin(int gridX, int gridY, int value) {
+    public Coin(int gridX, int gridY, double value) {
         this.gridX = gridX;
         this.gridY = gridY;
         this.value = value;
